tweaks(generic): Multiple different tweaks

+++ bukkit/src/main/java/dev/jsinco/brewery/bukkit/util/ColorUtil.java
@@ -74,7 +74,7 @@ public class ColorUtil {
     }
 
     // Returns a color closer to the destination color based on the interval and totalDuration
-    public static Color getNextColor(Color current, Color destination, long step, int duration) {
+    public static Color getNextColor(Color current, Color destination, long step, long duration) {
         float ratio = Math.min((float) step / (duration - 1), 1f);
         int red = (int) (ratio * destination.getRed() + (1f - ratio) * current.getRed());
         int green = (int) (ratio * destination.getGreen() + (1f - ratio) * current.getGreen());
